Add support for Heroku-20
This adds support for generating binaries for the upcoming Heroku-20
stack, and testing them in CI.

The Heroku-20 Dockerfile is based on that for Heroku-18, but with:
* the unused `python-{pip-whl,setuptools,wheel}` dependencies removed
* `python-pip` replaced with `python3-pip` since the former does not
  exist on Ubuntu 20.04.

The switch to Python 3 pip in the builder image only affects bob-builder,
since everything else (such as pip-diff) runs against the Python binaries
downloaded as part of the compile.

bob-builder has been updated to v0.0.18 to pick up the Python 3 fixes in:
https://github.com/heroku-python/bob-builder/pull/48

Before this is merged, the binaries will need to be built/uploaded using
the steps here:
https://github.com/heroku/heroku-buildpack-python/blob/master/builds/README.md

A decision will need to be made about whether to backfill older Python
versions, however for now we should at least provide the latest point
release of each supported Python branch:
https://devguide.python.org/#status-of-python-branches

Which are:
* `3.5.9`
* `3.6.10`
* `3.7.7`
* `3.8.2`

We'll also want to upload at least one version of PyPy, however the
current script is for the out of date `pypy3.6-7.2.0`, so we'll want to
generate newer binaries for all stacks, which is out of scope for now.

Refs W-7485877.